Abstracted out drawable stuff from TextBuffer
